+++ src/collective/solr/solr.py
@@ -34,6 +34,7 @@ from xml.etree.cElementTree import fromstring
 from xml.sax.saxutils import escape
 import codecs
 import urllib
+from collective.solr.exceptions import SolrConnectionException
 from collective.solr.interfaces import ISolrConnectionConfig
 from collective.solr.parser import SolrSchema
 from collective.solr.timeout import HTTPConnectionWithTimeout
@@ -42,23 +43,6 @@ from zope.component import queryUtility
 
 from logging import getLogger
 logger = getLogger(__name__)
-
-
-class SolrException(Exception):
-    """ An exception thrown by solr connections """
-
-    def __init__(self, httpcode='000', reason=None, body=None):
-        self.httpcode = httpcode
-        self.reason = reason
-        self.body = body
-
-    def __repr__(self):
-        return 'HTTP code=%s, Reason=%s, body=%s' % (
-            self.httpcode, self.reason, self.body
-        )
-
-    def __str__(self):
-        return 'HTTP code=%s, reason=%s' % (self.httpcode, self.reason)
 
 
 class SolrConnection:
@@ -110,7 +94,7 @@ class SolrConnection:
 
     def __errcheck(self, rsp):
         if rsp.status != 200:
-            ex = SolrException(rsp.status, rsp.reason)
+            ex = SolrConnectionException(rsp.status, rsp.reason)
             try:
                 ex.body = rsp.read()
             except:
@@ -161,7 +145,7 @@ class SolrConnection:
         for request in self.xmlbody:
             try:
                 responses.append(self.doSendXML(request))
-            except (SolrException, socket.error):
+            except (SolrConnectionException, socket.error):
                 logger.exception('exception during request %r', request)
             count += len(request)
         logger.debug(
@@ -187,7 +171,7 @@ class SolrConnection:
         status = parsed.attrib.get('status', 0)
         if status != 0:
             reason = parsed.documentElement.firstChild.nodeValue
-            raise SolrException(rsp.status, reason)
+            raise SolrConnectionException(rsp.status, reason)
         return parsed
 
     def escapeVal(self, val):
@@ -344,4 +328,4 @@ class SolrConnection:
                 xml = response.read()
                 return SolrSchema(xml.strip())
             self.__reconnect()          # force a new connection for each url
-        self.__errcheck(response)       # raise a solrexception
+        self.__errcheck(response)       # raise a SolrConnectionException